Use current timestamp as default

+++ mgof/mgof.py
@@ -7,11 +7,11 @@ class AnomalyDetector():
         self.r = redis.StrictRedis(host='localhost', port=6379)
 
     def _serialize_value(self, timestamp, value):
-        return "%d:%f" % (timestamp, value)
+        return "%f:%f" % (timestamp, value)
 
     def _read_value(self, string):
         ts, value = string.split(":")
-        return int(ts), float(value)
+        return float(ts), float(value)
 
     def post_metric(self, key, value, timestamp=None):
         timestamp = timestamp or time.time()
